+++ CHANGELOG.md
@@ -11,6 +11,7 @@
  * Update README files to include links for kite and `kite-spring-boot3-starter`
  * Enhance reflection utilities with entity detection and improve argument handling in `SqlParser`
  * Refactor `selectiveStrategy` type from function type to `Function1` for consistency
+ * Enhance SQL parameter resolution to support array element access and improve error handling
 
 ## 1.0.15